requirements.txt
# Core Libraries for Viral Clip Generation

# Audio Processing & Transcription
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy==1.11.4
numpy==1.24.3

# Video Processing & Generation
moviepy==1.0.3
Pillow==10.0.1
ffmpeg-python==0.2.0

# AI & Machine Learning
openai==1.3.7
torch==2.1.1
torchaudio==2.1.1
spacy==3.7.2

# Web Framework & API
Flask==3.0.0
FastAPI==0.104.1
uvicorn==0.24.0
flasgger==0.9.7.1
flask-cors==4.0.0

# Task Management & Background Processing
celery==5.3.4
redis==5.0.1

# External API Integrations
requests==2.31.0
httpx==0.25.2

# Utilities & Development
psutil==5.9.6
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
tenacity==8.2.3

# Additional dependencies
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
